refactor: restructure docs
To get the docs more in line with an Diátaxis (https://diataxis.fr)
approach, they need to be restructured to make them easier to approach
and use as a reference for users.

This commit restructures the documentation to start that process,
focusing only on moving *existing* documentation to new places, and
re-organizing the structure of the docs.

This work builds on Ben, Lachlan and Bailey's efforts to re-design and
restructure the documentation.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>